python/requirements.txt
scikit-learn==1.3.2
numpy==1.26.3
pandas==2.1.4
pymongo==4.6.1
python-dotenv==1.0.0